Remove `constval_array<>` in favor of `array_constant<>`
With the previous alias definition, an empty array, i.e.
`array_constant<T[]>`, could not be matched because it would comprise a
recursive pattern on an empty variadic pack.

Thinking ahead, with C++2a most types will be valid NTTPs, so
`array_constant<>` will become applicable for more types. Hence we now
wrap invalid NTTPs into an enum carrying the constval type. This permits
intuitive and homogeneous use of `array_constant<>` regardless of the
argument type. NTTP arguments must be converted to constvals using
`constval<>` (`constval14<>` in C++14), cf. the unit tests.

Starting with C++17 we do the same for `tuple_constant<>`. Tuples are
not normalized in C++14.