makeshift now is a static library
Admittedly the change was instigated by the DLL search path issues on
Windows; but there is little justification for having a dedicated DLL
for the handful of functions we export.

Perhaps we can also get rid of `MAKESHIFT_PUBLIC` etc. in the future? Or
is there a compelling reason to keep a shared library among our future
plans?